1.1 Update Update code, no functional changes Code clear

+++ VoiceReminder/ShowReminders.cs
@@ -1,14 +1,7 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
 using System.IO;
-using System.Linq;
 using System.Text;
-using System.Threading.Tasks;
 using System.Windows.Forms;
-using NAudio.Wave;
 using static VoiceReminder.ProgramDictionary;
 using static VoiceReminder.ProgramRegistry;
 
@@ -21,10 +14,12 @@ namespace VoiceReminder
         private int index;
         private string[] allfiles;
         private bool _isPlaying;
+        private bool _isFile;
         private WMPLib.WindowsMediaPlayer WMP;
         private string _lastRecordFile;
         private string _currentPath;
         private string _date;
+
         public ShowReminders()
         {
             InitializeComponent();
@@ -36,23 +31,32 @@ namespace VoiceReminder
             playVoice.Text = Words["playVoice"];
             deleteReminder.Text = Words["deleteReminder"];
             postponeReminder.Text = Words["postponeReminder"];
-            allfiles = Directory.GetFiles($@"{programDataPath}\TextData\", "*.*", SearchOption.AllDirectories);
+            allfiles = Directory.GetFiles($@"{ProgramDataPath}\TextData\", "*.*", SearchOption.AllDirectories);
             _isPlaying = false;
+            _isFile = false;
             index = 0;
             NewReminder();
         }
 
         private void deleteReminder_Click(object sender, EventArgs e)
         {
+            if (reminderName.Text.Equals("") && reminderText.Text.Equals(""))
+                return;
             File.Delete(_lastRecordFile);
             File.Delete(_currentPath);
-            allfiles = Directory.GetFiles($@"{programDataPath}\TextData\", "*.*", SearchOption.AllDirectories);
+            allfiles = Directory.GetFiles($@"{ProgramDataPath}\TextData\", "*.*", SearchOption.AllDirectories);
             index = 0;
             NewReminder();
         }
 
         private void playVoice_Click(object sender, EventArgs e)
         {
+            if (!_isFile)
+            {
+                MessageBox.Show(Words["noAudioRecorded"], Words["info"], MessageBoxButtons.OK, MessageBoxIcon.Information);
+                return;
+            }
+
             if (!_isPlaying)
             {
                 WMP = new WMPLib.WindowsMediaPlayer();
@@ -69,14 +73,15 @@ namespace VoiceReminder
 
         private void postponeReminder_Click(object sender, EventArgs e)
         {
+            if (reminderName.Text.Equals("") && reminderText.Text.Equals(""))
+                return;
             PostponeWindow postpone = new PostponeWindow();
             postpone.ShowDialog(allfiles[index]);
-
         }
 
         private void nextReminder_Click(object sender, EventArgs e)
         {
-            if (index != allfiles.Length-1)
+            if (index != allfiles.Length - 1)
                 index++;
             else
                 index = 0;
@@ -86,33 +91,45 @@ namespace VoiceReminder
 
         private void NewReminder()
         {
-            string[] fileFilling;
-            using (StreamReader sr = new StreamReader(allfiles[index]))
+            try
             {
-                fileFilling = sr.ReadToEnd().Split('\n');
-            }
-
-            reminderName.Text = fileFilling[0];
-            _lastRecordFile = fileFilling[1];
-            _currentPath = allfiles[index];
-            var tmp = _currentPath.Split('\\');
-            _date = tmp[tmp.Length - 1].Replace(".txt", "");
-            int temp = 0;
-            int counter = 0;
-            for (int i = 0; i < _date.Length; i++)
-            {
-                if (_date[i] == '_')
-                    counter++;
-                if (counter == 2)
+                var fileFilling = File.ReadAllLines(allfiles[index], Encoding.GetEncoding(1251));
+                reminderName.Text = fileFilling[0];
+                _lastRecordFile = fileFilling[1];
+                _currentPath = allfiles[index];
+                var tmp = _currentPath.Split('\\');
+                _date = tmp[tmp.Length - 1].Replace(".txt", "");
+                int temp = 0;
+                int counter = 0;
+                for (int i = 0; i < _date.Length; i++)
                 {
-                    temp = i;
-                    break;
+                    if (_date[i] == '_')
+                        counter++;
+                    if (counter == 2)
+                    {
+                        temp = i;
+                        break;
+                    }
                 }
+
+                _date = _date.Remove(0, temp + 1);
+                var partsDate = _date.Split('_');
+                partsDate[1] = partsDate[1].Replace('.', ':');
+                reminderText.Text = $"{Words["reminderIs"]} {partsDate[0]} {partsDate[1]}";
+                if (_lastRecordFile.Equals("0"))
+                {
+                    _isFile = false;
+                    return;
+                }
+
+                _isFile = true;
             }
-            _date = _date.Remove(0, temp + 1);
-            var partsDate = _date.Split('_');
-            partsDate[1] = partsDate[1].Replace('.', ':');
-            reminderText.Text = $"{Words["reminderIs"]} {partsDate[0]} {partsDate[1]}";
+            catch (IndexOutOfRangeException)
+            {
+                reminderName.Text = "";
+                reminderText.Text = "";
+                _lastRecordFile = "0";
+            }
         }
 
         private void previousReminder_Click(object sender, EventArgs e)
@@ -120,7 +137,7 @@ namespace VoiceReminder
             if (index != 0)
                 index--;
             else
-                index = allfiles.Length-1;
+                index = allfiles.Length - 1;
 
             NewReminder();
         }
